Move test files to appropriate places
