+++ src/shared/logging/logBuffer.test.js
@@ -34,22 +34,35 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/* eslint-disable import/first */
+import createLogBuffer from './logBuffer';
 
-jest.mock('fs-extra', () => ({}));
+describe('logBuffer', () => {
+    let logBuffer;
 
-import { getNameFromNpmPackage } from '../fileUtil';
-
-describe('getNameFromNpmPackage', () => {
-    it('should return null if path is invalid', () => {
-        expect(getNameFromNpmPackage('(invalid)')).toBeNull();
+    beforeEach(() => {
+        logBuffer = createLogBuffer();
     });
 
-    it('should return null if file name has no no dash', () => {
-        expect(getNameFromNpmPackage('/path/to/package.tgz')).toBeNull();
+    it('should have no entries when buffer is empty', () => {
+        expect(logBuffer.size()).toEqual(0);
     });
 
-    it('should return name if file name is valid', () => {
-        expect(getNameFromNpmPackage('/path/to/my-package-1.2.3.tgz')).toEqual('my-package');
+    it('should have 1 entry when 1 entry has been added', () => {
+        logBuffer.addEntry({ id: 0 });
+        expect(logBuffer.size()).toEqual(1);
+    });
+
+    it('should return all added entries when clearing buffer', () => {
+        const inputEntries = [{ id: 0 }, { id: 1 }];
+        logBuffer.addEntry(inputEntries[0]);
+        logBuffer.addEntry(inputEntries[1]);
+        const outputEntries = logBuffer.clear();
+        expect(outputEntries).toEqual(inputEntries);
+    });
+
+    it('should have no entries when buffer has been cleared', () => {
+        logBuffer.addEntry({ id: 0 });
+        logBuffer.clear();
+        expect(logBuffer.size()).toEqual(0);
     });
 });
